requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
boto3==1.16.20
botocore==1.19.20
dj-database-url==0.5.0
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-js-asset==1.2.2
django-storages==1.10.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
reportlab==3.5.53
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
